Cleaned up the AG for Walrus and SC information.
Removed some content from the AG on Walrus and SC storage
that was taken from the wiki but is too low-level for
inclusion in the AG. Specifically, removed sections that
detailed the mechanisms for handling volume operations as
those are subject to change between minor and subminor
versions.

Also removed and reworked some of the Walrus section on
filesystems and HA config of Walrus. Tried to make it a
bit more high-level such that it will not become inaccurate
on subsequent code updates.